1.105 Fabric and Forge Complete
Adds a Upgraded Furnace Block, this has 4 slots that will balance from first to last, stores the burn time. this also fixes the new block so it can cook items in last slot if others are empty. both dev env are now updated and and the same. still working on NeoForge