1.21.1-V1.143-B-001-Build
This adds the remote upgrade item, this is used to get the tick Accelerator to not need to be around the blocks you are ticking up. you can also set the range using efficacy upgrades. Grinders now will give you a 33% chance to triple your items with efficacy upgrade in